feat: refactor health check to allow spaces in SHELLBOT
In ef73bd45e07dc9669b20d97aa4b06200e7ea11f3 we added a support for
a command-line argument that can be used to override the default
prompt, but actually supplying an arguments means that the old
`vim.fn.executable` check won't work.

The simplest fix is to just split on whitespace and check the first
item to see if it's executable. This isn't perfect, because if
somebody installed the executable at a path like `/My/Path With
Spaces/bin/shellbot`, then we'd test `/My/Path` and report a false
negative. But for the common case, this is still an improvement.